topup-backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LangChain and LangGraph
langchain==0.1.6
langchain-openai==0.0.5
langgraph==0.0.20

# Vector DB and embeddings
chromadb==0.4.22

# Data validation and models
pydantic==2.5.3
pydantic-settings==2.1.0

# Data processing
pandas==2.2.0
numpy==1.26.3

# Database
sqlite-utils==3.36

# Visualization
plotly==5.18.0

# OpenAI
openai==1.10.0

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.4
httpx==0.26.0
